+++ src/battle_anim_ghost.c
@@ -816,7 +816,7 @@ void AnimTask_DestinyBondWhiteShadow(u8 taskId)
 
 static void AnimTask_DestinyBondWhiteShadow_Step(u8 taskId)
 {
-    u16 i;
+    u32 i;
 
     struct Task *task = &gTasks[taskId];
 
@@ -1145,7 +1145,7 @@ void AnimTask_GrudgeFlames(u8 taskId)
 
 static void AnimTask_GrudgeFlames_Step(u8 taskId)
 {
-    u16 i;
+    u32 i;
     u8 spriteId;
     struct Task *task = &gTasks[taskId];
