+++ src/battle_anim_throw.c
@@ -1765,7 +1765,7 @@ static void PokeBallOpenParticleAnimation_Step2(struct Sprite *sprite)
 
 static void TimerBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x = gTasks[taskId].data[1];
     u8 y = gTasks[taskId].data[2];
     u8 subpriority = gTasks[taskId].data[4];
@@ -1794,7 +1794,7 @@ static void TimerBallOpenParticleAnimation(u8 taskId)
 
 static void DiveBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x = gTasks[taskId].data[1];
     u8 y = gTasks[taskId].data[2];
     u8 subpriority = gTasks[taskId].data[4];
@@ -1824,7 +1824,7 @@ static void DiveBallOpenParticleAnimation(u8 taskId)
 // Also used for Net Ball
 static void SafariBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x = gTasks[taskId].data[1];
     u8 y = gTasks[taskId].data[2];
     u8 subpriority = gTasks[taskId].data[4];
@@ -1854,7 +1854,7 @@ static void SafariBallOpenParticleAnimation(u8 taskId)
 // Also used for Nest Ball
 static void UltraBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x = gTasks[taskId].data[1];
     u8 y = gTasks[taskId].data[2];
     u8 subpriority = gTasks[taskId].data[4];
@@ -1890,7 +1890,7 @@ static void GreatBallOpenParticleAnimation(u8 taskId)
     }
     else
     {
-        u8 i;
+        u32 i;
         u8 x = gTasks[taskId].data[1];
         u8 y = gTasks[taskId].data[2];
         u8 subpriority = gTasks[taskId].data[4];
@@ -1935,7 +1935,7 @@ static void FanOutBallOpenParticles_Step1(struct Sprite *sprite)
 
 static void RepeatBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x = gTasks[taskId].data[1];
     u8 y = gTasks[taskId].data[2];
     u8 subpriority = gTasks[taskId].data[4];
@@ -1972,7 +1972,7 @@ static void RepeatBallOpenParticleAnimation_Step1(struct Sprite *sprite)
 
 static void MasterBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i, j;
+    u32 i, j;
     u8 x = gTasks[taskId].data[1];
     u8 y = gTasks[taskId].data[2];
     u8 subpriority = gTasks[taskId].data[4];
@@ -2013,7 +2013,7 @@ static void MasterBallOpenParticleAnimation(u8 taskId)
 
 static void PremierBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x = gTasks[taskId].data[1];
     u8 y = gTasks[taskId].data[2];
     u8 subpriority = gTasks[taskId].data[4];
@@ -2150,7 +2150,7 @@ static void MoonBallParticle_Step(struct Sprite *sprite)
 
 static void FriendBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x = gTasks[taskId].data[1];
     u8 y = gTasks[taskId].data[2];
     u8 subpriority = gTasks[taskId].data[4];
@@ -2202,7 +2202,7 @@ static void FriendBallParticle_Step(struct Sprite *sprite)
 
 static void FastBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x = gTasks[taskId].data[1];
     u8 y = gTasks[taskId].data[2];
     u8 subpriority = gTasks[taskId].data[4];
