+++ src/battle_controller_recorded_opponent.c
@@ -823,7 +823,7 @@ static void RecordedOpponentEndExecution(void)
 static void RecordedOpponentHandleSetMonData(void)
 {
     u8 monToCheck;
-    u8 i;
+    u32 i;
 
     if (gBattleBufferA[gActiveBattler][2])
     {
@@ -1061,7 +1061,7 @@ static void SetRecordedOpponentMonData(u8 monId)
 static void RecordedOpponentHandleSetRawMonData(void)
 {
     u8 *dst = (u8 *)&gEnemyParty[gBattlerPartyIndexes[gActiveBattler]] + gBattleBufferA[gActiveBattler][1];
-    u8 i;
+    u32 i;
 
     for (i = 0; i < gBattleBufferA[gActiveBattler][2]; i++)
         dst[i] = gBattleBufferA[gActiveBattler][3 + i];
