+++ src/battle_factory_screen.c
@@ -1254,7 +1254,7 @@ static void CB2_InitSelectScreen(void)
 
 static void Select_InitMonsData(void)
 {
-    u8 i;
+    u32 i;
 
     if (sFactorySelectScreen != NULL)
         return;
@@ -1300,7 +1300,7 @@ static void Select_InitAllSprites(void)
 
 static void Select_DestroyAllSprites(void)
 {
-    u8 i;
+    u32 i;
 
     for (i = 0; i < SELECTABLE_MONS_COUNT; i++)
         DestroySprite(&gSprites[sFactorySelectScreen->mons[i].ballSpriteId]);
@@ -1421,7 +1421,7 @@ static void Select_SetBallSpritePaletteNum(u8 id)
 
 static void Select_Task_OpenSummaryScreen(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 currMonId;
 
     switch (gTasks[taskId].tState)
@@ -1746,7 +1746,7 @@ static void CreateFrontierFactorySelectableMons(u8 firstMonId)
 
 static void CreateSlateportTentSelectableMons(u8 firstMonId)
 {
-    u8 i, j;
+    u32 i, j;
     u8 ivs = 0;
     u8 level = 30;
     u8 friendship = 0;
@@ -1777,7 +1777,7 @@ static void CreateSlateportTentSelectableMons(u8 firstMonId)
 
 static void Select_CopyMonsToPlayerParty(void)
 {
-    u8 i, j;
+    u32 i, j;
 
     for (i = 0; i < FRONTIER_PARTY_SIZE; i++)
     {
@@ -2022,7 +2022,7 @@ static void Select_ReshowMonSprite(void)
 
 static void Select_CreateChosenMonsSprites(void)
 {
-    u8 i, j;
+    u32 i, j;
 
     for (i = 0; i < FRONTIER_PARTY_SIZE; i++)
     {
@@ -2218,7 +2218,7 @@ static void Select_SetWinRegs(s16 mWin0H, s16 nWin0H, s16 mWin0V, s16 nWin0V)
 
 static bool32 Select_AreSpeciesValid(u16 monId)
 {
-    u8 i, j;
+    u32 i, j;
     u32 species = gFacilityTrainerMons[monId].species;
     u8 selectState = sFactorySelectScreen->selectingMonsState;
 
@@ -2750,7 +2750,7 @@ static void Swap_Task_FadeOutSpeciesName(u8 taskId)
 #define tBallCycled(i) data[(i) + 1]
 static void Swap_Task_SlideCycleBalls(u8 taskId)
 {
-    s8 i;
+    s32 i;
     u8 lastX;
     bool8 finished;
 
@@ -2843,7 +2843,7 @@ static void Swap_Task_SlideCycleBalls(u8 taskId)
 
 static void Swap_Task_SlideButtonOnOffScreen(u8 taskId)
 {
-    u8 i, j;
+    u32 i, j;
     s32 posX = 0;
     s8 deltaX = gTasks[taskId].tXIncrement;
     bool8 sliding;
@@ -3159,7 +3159,7 @@ static void Swap_Task_ScreenInfoTransitionIn(u8 taskId)
 // For switching between the swap screens with the player's / enemy's parties
 static void Swap_Task_SwitchPartyScreen(u8 taskId)
 {
-    u8 i;
+    u32 i;
     if (sFactorySwapScreen->monPicAnimating == TRUE)
         return;
 
@@ -3392,7 +3392,7 @@ static void CB2_InitSwapScreen(void)
 
 static void Swap_InitAllSprites(void)
 {
-    u8 i;
+    u32 i;
     u8 x;
     struct SpriteTemplate spriteTemplate;
 
@@ -3486,7 +3486,7 @@ static void Swap_InitAllSprites(void)
 
 static void Swap_DestroyAllSprites(void)
 {
-    u8 i, j;
+    u32 i, j;
 
     for (i = 0; i < FRONTIER_PARTY_SIZE; i++)
         DestroySprite(&gSprites[sFactorySwapScreen->ballSpriteIds[i]]);
@@ -3617,7 +3617,7 @@ static void Swap_UpdateMenuCursorPosition(s8 direction)
 
 static void Swap_HighlightActionButton(u8 actionId)
 {
-    u8 i;
+    u32 i;
 
     for (i = 0; i < ARRAY_COUNT(sFactorySwapScreen->pkmnForSwapButtonSpriteIds[0]); i++)
     {
@@ -3644,7 +3644,7 @@ static void Swap_HighlightActionButton(u8 actionId)
 
 static void Swap_HideActionButtonHighlights(void)
 {
-    u8 i;
+    u32 i;
 
     for (i = 0; i < ARRAY_COUNT(sFactorySwapScreen->pkmnForSwapButtonSpriteIds[0]); i++)
     {
@@ -4113,7 +4113,7 @@ static void Swap_TaskCantHaveSameMons(u8 taskId)
 
 static bool8 Swap_AlreadyHasSameSpecies(u8 monId)
 {
-    u8 i;
+    u32 i;
     u16 species = GetMonData(&gEnemyParty[monId], MON_DATA_SPECIES, NULL);
 
     for (i = 0; i < FRONTIER_PARTY_SIZE; i++)
