+++ src/easy_chat.c
@@ -1869,7 +1869,7 @@ static void SetSelectedWord(u16 easyChatWord)
 // Compare current phrase to the original saved phrase
 static bool8 DidPhraseChange(void)
 {
-    u16 i;
+    u32 i;
     for (i = 0; i < sEasyChatScreen->maxWords; i++)
     {
         if (sEasyChatScreen->currentPhrase[i] != sEasyChatScreen->savedPhrase[i])
@@ -2362,7 +2362,7 @@ static int FooterHasFourOptions_(void)
 
 static bool8 IsPhraseDifferentThanPlayerInput(const u16 *phrase, u8 phraseLength)
 {
-    u8 i;
+    u32 i;
 
     for (i = 0; i < phraseLength; i++)
     {
@@ -2659,7 +2659,7 @@ static bool8 ReprintPhrase(void)
 
 static bool8 UpdateMainCursor(void)
 {
-    u8 i;
+    u32 i;
     u16 *currentPhrase;
     u16 *ecWord;
     u8 frameId;
@@ -4463,7 +4463,7 @@ u16 EasyChat_GetNumWordsInGroup(u8 groupId)
 
 static bool8 IsEasyChatWordInvalid(u16 easyChatWord)
 {
-    u16 i;
+    u32 i;
     u8 groupId;
     u32 index;
     u16 numWords;
@@ -4537,7 +4537,7 @@ u8 *CopyEasyChatWord(u8 *dest, u16 easyChatWord)
 
 u8 *ConvertEasyChatWordsToString(u8 *dest, const u16 *src, u16 columns, u16 rows)
 {
-    u16 i, j;
+    u32 i, j;
     u16 numColumns = columns - 1;
 
     for (i = 0; i < rows; i++)
@@ -4583,7 +4583,7 @@ static u16 GetEasyChatWordStringLength(u16 easyChatWord)
 
 static bool8 CanPhraseFitInXRowsYCols(const u16 *easyChatWords, u8 numRows, u8 numColumns, u16 maxLength)
 {
-    u8 i, j;
+    u32 i, j;
 
     for (i = 0; i < numColumns; i++)
     {
@@ -4692,7 +4692,7 @@ void UnlockAdditionalPhrase(u8 additionalPhraseId)
 
 static u8 GetNumAdditionalPhrasesUnlocked(void)
 {
-    u8 i;
+    u32 i;
     u8 numAdditionalPhrasesUnlocked;
 
     for (i = 0, numAdditionalPhrasesUnlocked = 0; i < NUM_ADDITIONAL_PHRASES; i++)
@@ -4706,7 +4706,7 @@ static u8 GetNumAdditionalPhrasesUnlocked(void)
 
 u16 GetNewHipsterPhraseToTeach(void)
 {
-    u16 i;
+    u32 i;
     u16 additionalPhraseId;
     u8 numAdditionalPhrasesUnlocked = GetNumAdditionalPhrasesUnlocked();
     if (numAdditionalPhrasesUnlocked == NUM_ADDITIONAL_PHRASES)
@@ -4739,7 +4739,7 @@ static bool8 EasyChatIsNationalPokedexEnabled(void)
 
 static u16 GetRandomUnlockedEasyChatPokemon(void)
 {
-    u16 i;
+    u32 i;
     u16 numWords;
     const u16 *species;
     u16 index = EasyChat_GetNumWordsInGroup(EC_GROUP_POKEMON);
@@ -4768,7 +4768,7 @@ static u16 GetRandomUnlockedEasyChatPokemon(void)
 
 void InitEasyChatPhrases(void)
 {
-    u16 i, j;
+    u32 i, j;
 
     for (i = 0; i < ARRAY_COUNT(sDefaultProfileWords); i++)
         gSaveBlock1Ptr->easyChatProfile[i] = sDefaultProfileWords[i];
@@ -4855,7 +4855,7 @@ static const u8 *GetEasyChatWordGroupName(u8 groupId)
 
 static u8 *CopyEasyChatWordPadded(u8 *dest, u16 easyChatWord, u16 totalChars)
 {
-    u16 i;
+    u32 i;
     u8 *str = CopyEasyChatWord(dest, easyChatWord);
     for (i = str - dest; i < totalChars; i++)
     {
@@ -4967,7 +4967,7 @@ static u16 SetSelectedWordGroup_GroupMode(u16 groupId)
 
 static u16 SetSelectedWordGroup_AlphabetMode(u16 groupId)
 {
-    u16 i;
+    u32 i;
     u16 totalWords;
 
     for (i = 0, totalWords = 0; i < sWordData->numUnlockedAlphabetWords[groupId]; i++)
@@ -5034,7 +5034,7 @@ static u8 IsEasyChatWordUnlocked(u16 easyChatWord)
 
 void InitializeEasyChatWordArray(u16 *words, u16 length)
 {
-    u16 i;
+    u32 i;
     for (i = length - 1; i != EC_EMPTY_WORD; i--)
         *(words++) = EC_EMPTY_WORD;
 }
